API

UCX.AMHandler — Type

AMHandler(func)

Arguments to callback

worker::UCXWorker

data::Ptr{Cvoid}

length::Csize_t

reply_ep::API.ucp_ep_h

flags::Cuint

Return values

The callback func needs to return either UCX.API.UCS_OK or UCX.API.UCS_INPROGRESS. If it returns UCX.API.UCS_INPROGRESS it must call am_data_release(worker, data), or call am_recv.

UCX.progress — Function

progress(worker::UCXWorker)

Allows worker to make progress, this includes finalizing requests and call callbacks.

Returns true if progress was made, false if no work was waiting.

UCX.set_library — Function

set_library(libucp=nothing)

Configure UCX.jl to use a specific UCX library by saving the absolute path to libucp as a preference. Calling the function with no arguments will reset the preference to use UCX from the Yggdrasil-provided UCX_jll. This is a compile-time preference, so Julia must be restarted for the change to take effect.